refactoring limits into a package
